refactor(HS): batch EOS derivs + split binding vs exploratory tests (#2999)
Addresses Ian's review:

- hs_leg_departure (active production code): the per-term
  calc_alpha0_deriv_nocache / calc_alphar_deriv_nocache calls each
  compute the FULL HelmholtzDerivatives set and return one component, so
  the lambda-continuation residual recomputed everything 6-9x per
  evaluation. Get the ideal-gas and residual blocks once each via
  calc_all_alpha0_derivs_nocache + residual_helmholtz->all(). Same math
  (residual all() is exactly what calc_alphar_deriv_nocache calls
  internally; the alpha0 reducing-state scaling is identity here), now
  one evaluation. Same fix applied to the prototype's lprops.

- Split CoolProp-Tests-HS.cpp so review can focus on the active code:
  the binding regression test (HS_prod2ph, the only default-suite test,
  exercises the shipped HmolarSmolar_INPUTS flash) stays here; the
  exploratory prototype solvers (continue_HS / solve_HS_continuation /
  _isentrope / _departure / _cascade) and their [.]-hidden A/B
  characterization, benchmark and probe tests move to the new
  CoolProp-Tests-HS-prototypes.cpp. No test logic changed; verified
  [HS_prod2ph], [HS_dep] (6396/6396 round-trips) and [HS_casc] pass.
